services/app/requirements.txt
flask==2.2.3
python-dotenv==1.0.0
blinker==1.5
flask-dance==6.2.0 
flask-login==0.6.2
flask-cors==3.0.10 
flask-marshmallow==0.14.0 
flask-migrate==4.0.4 
flask-sqlalchemy==3.0.3
gunicorn==20.1.0
marshmallow-sqlalchemy==0.29.0 
psycopg2-binary==2.9.5
sqlalchemy-utils==0.40.0